Views/Helpers/Main: implement login and signup.
This is actually the meat and potatoes of this branch. I'm actually
looking forward to also implement firestore and a cloud experience for
the user.
The signup screen is a little janky and will certainly need some
polishing/a rewrite in the future, but it will do for now, it allows the
user to sign up and see local files.
Those local files will be moved to cloud ASAP, and then I guess this
branch will be considered done, as this is the minimum firebase
integration we need at the moment.
Other changes might be reducing BLoC calls and local caching, as right
now we're on the Spark plan and we're all outta mone to pay for Firebase.