Bloc: implement a Firestore Bloc.
This will be used to fetch/write user data.
Data is not cached as of this commit, but caching techniques need to be
used in order ot reduce read count, therefore saving us money in the
long run.

Widgets can use this bloc through a BlocBuilder, listening to
FirestoreEvent(s) for read/write states.

Right now, only read states exist, but this will obviously change in the
future.